feat: lint-cleanup-website-004 - Preserve visible workflow and current-selection behavior after decomposition
