+++ docs/internal/processes/development-guide-relevant-files.md
@@ -50,6 +50,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/features/current-selection/current-selection-widget.tsx` and `ui/src/features/current-selection/provider-session-attempts.tsx` | Current-selection locale threading seam | Localized current-selection surfaces should accept an optional `locale` prop at the widget seam and forward only the narrow message-derived labels each detail surface needs, so feature-local catalogs can localize workstation or request cards without inventing a dashboard-wide locale store or coupling shared drill-down helpers to one message package. |
 | `ui/src/features/workflow-activity/messages/` | Feature-owned workflow activity catalogs | Workflow-activity localization should keep dashboard flow legend and graph-import copy in this adjacent feature package, using the shared `ui/src/i18n` resolver and direct lookup tests to prove supported locales, fallback behavior, and function-valued accessibility labels instead of hardcoding visible or accessible strings inside the React Flow host or dialog shell; when a catalog owns icon-label patterns, drag/drop status helpers, or error-copy helpers, test at least one non-default locale invocation so untranslated helper branches do not hide in CI. |
 | `ui/src/features/workflow-activity/react-flow-current-activity-card.tsx` | Workflow-activity locale boundary | Dashboard graph hosts should accept an optional `locale` prop and derive localized default legend, graph-import overlay, import error, and preview-dialog shell copy at the workflow-activity seam, so future localized routes or focused stories can exercise non-default graph copy without teaching React Flow primitives, topology builders, or unrelated dashboard widgets about the catalog; when that host seam owns locale fallback, add rendered integration coverage there for unsupported-locale fallback instead of relying only on message-catalog lookup tests, and keep translated validation alerts, dismiss/reset behavior, and preview-dialog activation-failure semantics covered at this host seam. |
+| `ui/src/features/workflow-activity/react-flow-current-activity-card-view-model.ts` | Workflow-activity graph-state seam | Keep topology-layout caching, persisted node positions, and React Flow edge/node derivation behind this feature-local hook seam so the rendered graph host can stay focused on visible states, drag/drop affordances, and dialog wiring while graph-state regressions still fail through the existing rendered card tests and stories. |
 | `ui/src/features/header/dashboard-export-dialog.tsx` | Dashboard-to-export locale boundary | Dashboard-owned export hosts should accept an optional `locale` prop and thread it into the feature dialog seam so stories, focused tests, and future localized routes can exercise non-default export copy without teaching shared dialog primitives or global app shells about feature-owned catalogs. |
 | `ui/src/features/header/dashboard-header.tsx` | Dashboard export-trigger locale boundary | Dashboard-owned controls that open a localized feature dialog should resolve their accessible names from the same feature-owned catalog, with an optional `locale` prop at the header seam for focused tests and future localized routes instead of duplicating hardcoded trigger copy beside the dialog. |
 | `ui/src/features/header/tick-slider-control.tsx` | Dashboard timeline locale boundary | Timeline slider controls should accept the same optional `locale` prop as the owning header and resolve visible plus accessible tick copy through the header feature catalog, keeping interpolation local to the control instead of recreating hardcoded tick-status strings in parent shells or tests. |
@@ -135,6 +136,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/features/timeline/state/factoryTimelineStore.test.ts` | Focused timeline replay contract surface | Replay cleanup work should prove canonical and supported legacy dispatch request or response behavior here by asserting reconstructed workstation requests, trace dispatches, runtime sessions, and world state rather than source-layout details. |
 | `ui/src/App.current-selection.test.tsx` | Dedicated current-selection app regression surface | App-level selected-work, trace-card, responsive-layout, terminal-state, and mixed inference/script dispatch-history regressions should live here once the focused current-selection flow would otherwise push `ui/src/App.test.tsx` past the repository’s file or function size limits. |
 | `ui/src/features/current-selection/workstation-request-detail.tsx` | Workstation-request detail fallback seam | Request-detail rendering, script-versus-inference alternate panels, and sparse projected field fallbacks should stay centralized here so current-selection request diagnostics keep one observable detail contract instead of splitting fallback copy across summary cards or app shells. |
+| `ui/src/features/current-selection/workstation-request-detail-response.tsx` and `ui/src/features/current-selection/workstation-request-detail-view.ts` | Workstation-request alternate-response seam | When request diagnostics need more detail without growing the top-level card past the lint-size boundary, keep the response/failure sections and their derived view state in these focused helpers so the top-level request card remains a thin composition layer and the observable fallback copy continues to be proven by the existing rendered request-detail tests. |
 | `ui/src/features/current-selection/workstation-request-detail.test.tsx` | Workstation-request detail regression surface | Customer-visible request-detail regressions should assert rendered markdown/plain-text prompt handling plus sparse request, trace, script-request, and script-response fallback copy here rather than relying on helper-only tests or broader app smoke coverage. |
 | `ui/src/features/current-selection/work-item-card.tsx` | Selected-work summary ownership seam | The selected-work card summary should stay limited to stable work metadata, relationships, and dispatch count here; request, response, provider-session, and failure diagnostics for specific dispatches should not be reintroduced into this top-level summary surface. |
 | `ui/src/features/current-selection/work-item-card.test.tsx` | Selected-work component regression surface | Selected-work tests should prove the relationship-graph surface plus dispatch-history-owned request, response, nested attempt, related-work selection, and trace-link behavior here so simplification regressions fail at the component seam before broader app flows are involved. |
