+++ ui/src/features/workflow-activity/react-flow-current-activity-card-types.ts
@@ -1,0 +1,33 @@
+import type {
+  DashboardActiveExecution,
+  DashboardSnapshot,
+  DashboardWorkItemRef,
+} from "../../api/dashboard/types";
+import type { FactoryValue } from "../../api/named-factory";
+import type { FactoryPngImportValue, ReadFactoryImportFile } from "../import";
+import type { CurrentActivityImportController } from "./current-activity-import-controller";
+
+export type CurrentActivitySelection =
+  | { kind: "node"; nodeId: string }
+  | { kind: "state-node"; placeId: string }
+  | { kind: "work-item"; dispatchId: string; nodeId: string; workID: string };
+
+export interface ReactFlowCurrentActivityCardProps {
+  activateFactory?: (value: FactoryValue) => Promise<FactoryValue>;
+  importController?: CurrentActivityImportController;
+  locale?: string;
+  now: number;
+  onFactoryActivated?: () => void;
+  onFactoryImportReady?: (value: FactoryPngImportValue, file: File) => void;
+  onSelectStateNode: (placeId: string) => void;
+  onSelectWorkItem: (
+    dispatchId: string,
+    nodeId: string,
+    execution: DashboardActiveExecution,
+    workItem: DashboardWorkItemRef,
+  ) => void;
+  onSelectWorkstation: (nodeId: string) => void;
+  readFactoryImportFile?: ReadFactoryImportFile;
+  selection: CurrentActivitySelection | null;
+  snapshot: DashboardSnapshot;
+}
