+++ ui/src/features/workflow-activity/react-flow-current-activity-card.tsx
@@ -1,73 +1,40 @@
 import "@xyflow/react/dist/style.css";
 
-import {
-  applyNodeChanges,
-  Background,
-  Controls,
-  type FitViewOptions,
-  type NodeChange,
-  ReactFlow,
-} from "@xyflow/react";
+import { Background, Controls, ReactFlow } from "@xyflow/react";
 import type { CSSProperties } from "react";
-import { useCallback, useEffect, useMemo, useState } from "react";
 
-import type {
-  DashboardActiveExecution,
-  DashboardSnapshot,
-  DashboardWorkItemRef,
-} from "../../api/dashboard/types";
-import type { FactoryValue } from "../../api/named-factory";
 import { DASHBOARD_SECTION_HEADING_CLASS } from "../../components/ui/dashboard-typography";
 import { cx } from "../../lib/cx";
-import {
-  CURRENT_ACTIVITY_NODE_TYPES,
-  type CurrentActivityNode,
-} from "../flowchart/current-activity-nodes";
-import { buildGraphLayout, type GraphLayout } from "../flowchart/layout";
-import {
-  FactoryImportPreviewDialog,
-  type FactoryPngImportValue,
-  type ReadFactoryImportFile,
-} from "../import";
-import {
-  type CurrentActivityImportController,
-  useCurrentActivityImportController,
-} from "./current-activity-import-controller";
+import { FactoryImportPreviewDialog } from "../import";
+import { useCurrentActivityImportController } from "./current-activity-import-controller";
 import {
   DashboardFlowAxisLegend,
   getDefaultDashboardFlowAxisLegendEdgeItems,
   getDefaultDashboardFlowAxisLegendIconItems,
 } from "./dashboard-flow-axis-legend";
-import {
-  groupActiveExecutionsByWorkstationNodeID,
-  useActiveExecutions,
-} from "./react-flow-current-activity-card-active-executions";
-import {
-  buildActiveGraphHighlights,
-  buildActiveItemLabelsByPlaceId,
-  buildCurrentActivityNodes,
-  buildGraphEdges,
-  buildHandleAssignments,
-  buildVisibleGraphEdges,
-  EMPTY_GRAPH_LAYOUT,
-  EMPTY_NODE_POSITIONS,
-  initialFocusNodes,
-} from "./react-flow-current-activity-card-graph";
+import { useCurrentActivityGraphViewModel } from "./react-flow-current-activity-card-view-model";
 import {
   GraphDropOverlay,
   GraphImportErrorPanel,
   graphDropStateAttribute,
 } from "./react-flow-current-activity-card-import";
-import { currentActivityGraphKey, currentActivityTopologyKey } from "./react-flow-current-activity-card-keys";
-import { useCurrentActivityGraphStore } from "./state/currentActivityGraphStore";
+import type { ReactFlowCurrentActivityCardProps } from "./react-flow-current-activity-card-types";
 
-export { currentActivityGraphKey, currentActivityTopologyKey } from "./react-flow-current-activity-card-keys";
+export {
+  currentActivityGraphKey,
+  currentActivityTopologyKey,
+} from "./react-flow-current-activity-card-keys";
+export type {
+  CurrentActivitySelection,
+  ReactFlowCurrentActivityCardProps,
+} from "./react-flow-current-activity-card-types";
 
 const GRAPH_BACKGROUND_COLOR = "var(--color-af-edge-muted-soft)";
 const GRAPH_BACKGROUND_GAP = 24;
 const GRAPH_BACKGROUND_SIZE = 1;
 
-type CSSPropertiesWithVariables = CSSProperties & Record<`--${string}`, string | number>;
+type CSSPropertiesWithVariables = CSSProperties &
+  Record<`--${string}`, string | number>;
 
 const GRAPH_CONTROLS_STYLE: CSSPropertiesWithVariables = {
   "--xy-controls-box-shadow": "none",
@@ -86,8 +53,6 @@ const GRAPH_CONTROLS_STYLE: CSSPropertiesWithVariables = {
   overflow: "hidden",
 };
 
-const GRAPH_LAYOUT_CACHE = new Map<string, GraphLayout>();
-const GRAPH_LAYOUT_PROMISE_CACHE = new Map<string, Promise<GraphLayout>>();
 const CURRENT_ACTIVITY_CARD_CLASS =
   "relative flex h-full min-h-0 min-w-0 flex-col rounded-3xl border border-af-overlay/10 bg-af-surface/72 p-[1.2rem] shadow-af-panel backdrop-blur-[18px] max-[720px]:p-4";
 const CURRENT_ACTIVITY_HEADER_CLASS = "mb-4";
@@ -96,11 +61,6 @@ const CURRENT_ACTIVITY_EYEBROW_CLASS =
 const CURRENT_ACTIVITY_LEGEND_CLASS =
   "absolute left-7 top-7 z-10 max-[720px]:left-4 max-[720px]:right-4 max-[720px]:top-4";
 const CURRENT_ACTIVITY_TITLE_CLASS = cx("m-0", DASHBOARD_SECTION_HEADING_CLASS);
-
-export type CurrentActivitySelection =
-  | { kind: "node"; nodeId: string }
-  | { kind: "state-node"; placeId: string }
-  | { kind: "work-item"; dispatchId: string; nodeId: string; workID: string };
 
 function CurrentActivityCardHeading() {
   return (
@@ -111,252 +71,6 @@ function CurrentActivityCardHeading() {
       </h2>
     </div>
   );
-}
-
-interface ReactFlowCurrentActivityCardProps {
-  activateFactory?: (value: FactoryValue) => Promise<FactoryValue>;
-  importController?: CurrentActivityImportController;
-  locale?: string;
-  now: number;
-  onFactoryActivated?: () => void;
-  onFactoryImportReady?: (value: FactoryPngImportValue, file: File) => void;
-  onSelectStateNode: (placeId: string) => void;
-  onSelectWorkItem: (
-    dispatchId: string,
-    nodeId: string,
-    execution: DashboardActiveExecution,
-    workItem: DashboardWorkItemRef,
-  ) => void;
-  onSelectWorkstation: (nodeId: string) => void;
-  readFactoryImportFile?: ReadFactoryImportFile;
-  selection: CurrentActivitySelection | null;
-  snapshot: DashboardSnapshot;
-}
-
-function useGraphLayout(snapshot: DashboardSnapshot) {
-  const topologyKey = useMemo(
-    () => currentActivityTopologyKey(snapshot.topology),
-    [snapshot.topology],
-  );
-  const layoutTopology = useMemo(() => snapshot.topology, [snapshot.topology]);
-  const [graphLayout, setGraphLayout] =
-    useState<GraphLayout>(EMPTY_GRAPH_LAYOUT);
-
-  useEffect(() => {
-    let cancelled = false;
-    const cachedLayout = GRAPH_LAYOUT_CACHE.get(topologyKey);
-    if (cachedLayout) {
-      setGraphLayout(cachedLayout);
-      return () => {
-        cancelled = true;
-      };
-    }
-
-    const inFlightLayout =
-      GRAPH_LAYOUT_PROMISE_CACHE.get(topologyKey) ??
-      buildGraphLayout(layoutTopology);
-    GRAPH_LAYOUT_PROMISE_CACHE.set(topologyKey, inFlightLayout);
-
-    inFlightLayout
-      .then((layout) => {
-        GRAPH_LAYOUT_CACHE.set(topologyKey, layout);
-        GRAPH_LAYOUT_PROMISE_CACHE.delete(topologyKey);
-        if (!cancelled) {
-          setGraphLayout(layout);
-        }
-      })
-      .catch(() => {
-        GRAPH_LAYOUT_PROMISE_CACHE.delete(topologyKey);
-        if (!cancelled) {
-          setGraphLayout(EMPTY_GRAPH_LAYOUT);
-        }
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [layoutTopology, topologyKey]);
-
-  return graphLayout;
-}
-
-function useCurrentActivityBaseNodes({
-  activeExecutionsByWorkstationNodeID,
-  activeGraphHighlights,
-  activeItemLabelsByPlaceId,
-  graphLayout,
-  handleAssignments,
-  now,
-  onSelectStateNode,
-  onSelectWorkItem,
-  onSelectWorkstation,
-  selection,
-  snapshot,
-  storedNodePositions,
-}: Pick<
-  ReactFlowCurrentActivityCardProps,
-  | "now"
-  | "onSelectStateNode"
-  | "onSelectWorkItem"
-  | "onSelectWorkstation"
-  | "selection"
-  | "snapshot"
-> & {
-  activeExecutionsByWorkstationNodeID: Record<
-    string,
-    DashboardActiveExecution[]
-  >;
-  activeGraphHighlights: ReturnType<typeof buildActiveGraphHighlights>;
-  activeItemLabelsByPlaceId: ReturnType<typeof buildActiveItemLabelsByPlaceId>;
-  graphLayout: GraphLayout;
-  handleAssignments: ReturnType<typeof buildHandleAssignments>;
-  storedNodePositions: typeof EMPTY_NODE_POSITIONS;
-}) {
-  return useMemo<CurrentActivityNode[]>(
-    () =>
-      buildCurrentActivityNodes({
-        activeExecutionsByWorkstationNodeID,
-        activeGraphHighlights,
-        activeItemLabelsByPlaceId,
-        graphLayout,
-        handleAssignments,
-        now,
-        onSelectStateNode,
-        onSelectWorkItem,
-        onSelectWorkstation,
-        selection,
-        snapshot,
-        storedNodePositions,
-      }),
-    [
-      activeExecutionsByWorkstationNodeID,
-      activeGraphHighlights,
-      activeItemLabelsByPlaceId,
-      graphLayout,
-      handleAssignments,
-      now,
-      onSelectStateNode,
-      onSelectWorkItem,
-      onSelectWorkstation,
-      selection,
-      snapshot,
-      storedNodePositions,
-    ],
-  );
-}
-
-function useCurrentActivityGraphViewModel({
-  now,
-  onSelectStateNode,
-  onSelectWorkItem,
-  onSelectWorkstation,
-  selection,
-  snapshot,
-}: Pick<
-  ReactFlowCurrentActivityCardProps,
-  | "now"
-  | "onSelectStateNode"
-  | "onSelectWorkItem"
-  | "onSelectWorkstation"
-  | "selection"
-  | "snapshot"
->) {
-  const activeExecutions = useActiveExecutions(snapshot);
-  const activeExecutionsByWorkstationNodeID = useMemo(
-    () => groupActiveExecutionsByWorkstationNodeID(activeExecutions),
-    [activeExecutions],
-  );
-  const graphLayout = useGraphLayout(snapshot);
-  const graphKey = useMemo(
-    () => currentActivityGraphKey(graphLayout),
-    [graphLayout],
-  );
-  const storedNodePositions = useCurrentActivityGraphStore(
-    (state) => state.positionsByGraphKey[graphKey] ?? EMPTY_NODE_POSITIONS,
-  );
-  const setStoredNodePosition = useCurrentActivityGraphStore(
-    (state) => state.setNodePosition,
-  );
-  const visibleGraphEdges = useMemo(
-    () => buildVisibleGraphEdges(graphLayout),
-    [graphLayout],
-  );
-  const handleAssignments = useMemo(
-    () => buildHandleAssignments(visibleGraphEdges),
-    [visibleGraphEdges],
-  );
-  const activeGraphHighlights = useMemo(
-    () => buildActiveGraphHighlights(activeExecutions, visibleGraphEdges),
-    [activeExecutions, visibleGraphEdges],
-  );
-  const activeItemLabelsByPlaceId = useMemo(
-    () => buildActiveItemLabelsByPlaceId(activeExecutions),
-    [activeExecutions],
-  );
-  const baseNodes = useCurrentActivityBaseNodes({
-    activeExecutionsByWorkstationNodeID,
-    activeGraphHighlights,
-    activeItemLabelsByPlaceId,
-    graphLayout,
-    handleAssignments,
-    now,
-    onSelectStateNode,
-    onSelectWorkItem,
-    onSelectWorkstation,
-    selection,
-    snapshot,
-    storedNodePositions,
-  });
-  const [nodes, setNodes] = useState<CurrentActivityNode[]>([]);
-
-  useEffect(() => {
-    setNodes((currentNodes) => {
-      const currentPositions = new Map(
-        currentNodes.map((node) => [node.id, node.position]),
-      );
-      return baseNodes.map((node) => ({
-        ...node,
-        position: currentPositions.get(node.id) ?? node.position,
-      }));
-    });
-  }, [baseNodes]);
-
-  const handleNodesChange = useCallback((changes: NodeChange[]) => {
-    setNodes(
-      (currentNodes) =>
-        applyNodeChanges(changes, currentNodes) as CurrentActivityNode[],
-    );
-  }, []);
-  const edges = useMemo(
-    () =>
-      buildGraphEdges(
-        activeGraphHighlights,
-        handleAssignments,
-        visibleGraphEdges,
-      ),
-    [activeGraphHighlights, handleAssignments, visibleGraphEdges],
-  );
-  const initialFitViewOptions = useMemo<FitViewOptions>(
-    () => ({
-      maxZoom: 1.15,
-      minZoom: 0.7,
-      nodes: initialFocusNodes(graphLayout),
-      padding: 0.18,
-    }),
-    [graphLayout],
-  );
-
-  return {
-    edges,
-    graphKey,
-    handleNodesChange,
-    initialFitViewKey:
-      initialFitViewOptions.nodes?.map((node) => node.id).join(":") ||
-      "full-graph",
-    initialFitViewOptions,
-    nodes,
-    setStoredNodePosition,
-  };
 }
 
 function EmptyCurrentActivityCard() {
@@ -442,7 +156,7 @@ export function ReactFlowCurrentActivityCard(
             key={graph.initialFitViewKey}
             maxZoom={2}
             minZoom={0.25}
-            nodeTypes={CURRENT_ACTIVITY_NODE_TYPES}
+            nodeTypes={graph.nodeTypes}
             nodes={graph.nodes}
             nodesDraggable={true}
             onNodeDragStop={(_, node) => {
